#132 code in scripts moved to sbpipe package
